Don't send Expect: 100-continue header
HTTP_Request2 will send an Expect: 100-continue header if the request body is greater than a certain size. If a request fails and is caught, the socket is in an unusable state and will time-out on future requests. Setting except to an empty string ensures that the request body is always sent on a request.