Fix getGuid() on max mt_rand
As identified by @znrk (#984), getGuid() is broken.

In fact, it's not only broken on the line identified for clock_seq_low,
but also in 2 other places. mt_rand() is inclusive for both the max and
the min, so mt_rand(0, 256) produces a max value of 0x100 rather than
0xff, overflowing from 2 bits to 3 bits. The fix is simple - changing
the max value on the 3 lines where it's broken to ensure that the
generated GUID is in line with the spec for version 4 (random GUID) at
<https://tools.ietf.org/html/rfc4122>.

This commit fixes #984.

I'll follow this commit up with a commit that introduces tests to
illustrate and cover this scenario.