Improve getGuid() tests
getGuid() is difficult to test well because of it's dependency on
mt_rand, which we can't control the behavior of in a test. By allowing
the consumer to pass in a random function to getGuid() (while still
defaulting to mt_rand), we increase the flexibility of the method and
also provide a way for a unit test to control the behavior of the
method.

There was a bug (#984) where getGuid() produced bad results when
mt_rand() generated a value at the max of it's possible range. By unit
testing it with values forced to the min and max, we help guarantee that
it produces a good result across the entire range (making sure that our
edge conditions are correct).