Upgrade to PHPUnit 5
Support for PHPUnit 4 ended on Feb 3, 2017. The transition from PHPUnit
4 to PHPUnit 5 is simple, and we should update the library. There's
little risk of breaking things since this is only a dev dependency.

In fact, support for PHPUnit 5 ended on Feb 2, 2018, but we can't
upgrade to PHPUnit 6 while still supporting PHP 5.6 (which will no
longer be supported after Jan 1, 2019).

The composer lockfile has also been updated. I ran this update in a php
5.6 environment so that it will not break for people who may still be
using that version of PHP.

<https://github.com/sebastianbergmann/phpunit/wiki/Release-Announcement-for-PHPUnit-5.0.0>
<https://phpunit.de/getting-started/phpunit-5.html>
<https://phpunit.de/getting-started/phpunit-4.html>
<https://secure.php.net/supported-versions.php>